alfred-devel: Add option to select download source
The current way to download the alfred master branch sources via a special script on
open-mesh.org doesn't work anymore because each https download requires the
checksum of this file. Thus the whole download implementation has to be
replaced to get everything via the native OpenWrt download helpers.

The supported methods are now:

* download via git
  - free selection of the git repository remote
  - free selection of the branch or tag
* copy from local source directory

Signed-off-by: Sven Eckelmann <sven@narfation.org>