alfred-devel: Suppress "error" about missing VERSION
The OpenWrt feeds script is trying to generate an index of packages and
then prints an error (actually warning) that this package doesn't have a
version.

The version of the package is not known yet because the source for the
package was not yet selected. To avoid that message, just set it to 0.0
when no other version information was specified.

Signed-off-by: Sven Eckelmann <sven@narfation.org>