batctl-devel: Provide different build variants
The batctl binary is currently optional for batman-adv installations. But
new configuration settings will only be exposed via generic netlink. The
batctl tool will therefore be required to modify them.

To also fit batctl in some of the smaller devices, a new variant tiny is
build which only provides the settings subcommands. The default variant is
equal to the old batctl package and full also provides the commands which
were disabled until now.

There are no actual variant package builds but only a configuration
settings. The actual openwrt-routing package requires variant package
builds but this would conflict with the per package build settings of the
batctl-devel package.

Signed-off-by: Sven Eckelmann <sven@narfation.org>